feat(refactor): Move StringError from authenticator to types module
